requirements.txt
asgiref==3.3.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudinary==1.25.0
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-cloudinary-storage==0.3.0
django-extra-views==0.13.0
django-haystack==3.0
django-heroku==0.3.1
django-oscar==3.0.2
django-phonenumber-field==3.0.1
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.8
factory-boy==2.12.0
Faker==6.4.1
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
idna==2.10
phonenumbers==8.12.18
Pillow==8.1.0
pip-autoremove==0.9.1
psycopg2==2.7.7
psycopg2-binary==2.8.6
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.1
whitenoise==5.2.0
